Siirretty lukon ottaminen scheduleriin

+++ valinta-tulos-service/src/main/scala/fi/vm/sade/valintatulosservice/sijoittelu/SijoitteluajoDeleteScheduler.scala
@@ -18,21 +18,32 @@ class SijoitteluajoDeleteScheduler(repository: DeleteSijoitteluRepository,
   val task = new Runnable {
     override def run(): Unit = Try {
 
-      val sijoitteluAjosToBeDestroyed = for(
-        (haku, size) <- repository.listHakuAndSijoitteluAjoCount().filter(_._2 > sijoitteluAjoLimit)
-      ) yield {
-        val sijoitteluAjoIds = repository.findSijoitteluAjotSkippingFirst(haku, sijoitteluAjoLimit)
+      if (!repository.acquireLockForSijoitteluajoCleaning(55).head) {
+        logger.info("sijoitteluajojen poistolukko: Lukkoa ei saatu. Joku muu VTS-node on vastuussa vanhojen sijoitteluajojen poistosta.")
+      } else {
+        logger.info("sijoitteluajojen poistolukko: Saatiin lukko. Tämä VTS-node on vastuussa vanhojen sijoitteluajojen poistosta. Suoritetaan siivous.")
 
-        (haku, sijoitteluAjoIds)
-      }
+        val sijoitteluAjosToBeDestroyed = for(
+          (haku, size) <- repository.listHakuAndSijoitteluAjoCount().filter(_._2 > sijoitteluAjoLimit)
+        ) yield {
+          val sijoitteluAjoIds = repository.findSijoitteluAjotSkippingFirst(haku, sijoitteluAjoLimit)
 
-      sijoitteluAjosToBeDestroyed.foreach {
-        case (haku, ids) =>
-          logger.info(s"Deleting ${ids.size} sijoitteluajos from $haku! ${ids.take(3)} ...")
-          Try(repository.deleteSijoitteluajot(haku, ids)) match {
-            case x if x.isSuccess => x
-            case Failure(e) => logger.error(s"Haun $haku sijoitteluajojen poisto epäonnistui", e)
-          }
+          (haku, sijoitteluAjoIds)
+        }
+
+        sijoitteluAjosToBeDestroyed.foreach {
+          case (haku, ids) =>
+            logger.info(s"Deleting ${ids.size} sijoitteluajos from $haku! ${ids.take(3)} ...")
+            Try(repository.deleteSijoitteluajot(haku, ids)) match {
+              case x if x.isSuccess => x
+              case Failure(e) => logger.error(s"Haun $haku sijoitteluajojen poisto epäonnistui", e)
+            }
+        }
+        if (repository.clearLockForSijoitteluajoCleaning(55).head) {
+          logger.info("sijoitteluajojen poistolukko: Sijoitteluajojen poisto on päättynyt ja lukitus onnistuneesti poistettu.")
+        } else {
+          logger.error("sijoitteluajojen poistolukko: jokin meni vikaan lukituksen poistossa.")
+        }
       }
     }
   }
